rename SessionsStore to SessionInfoStorage
