+++ doc/impl-thoughts/RECONNECTING.md
@@ -40,7 +40,7 @@ rooms should report how many messages they have queued up, and each time they se
     - put in own file
     - add waitFor (won't this leak if the promise never resolves?)
     - decide whether we want to inherit (no?)
- - cleanup Reconnector with recent changes, move generic code, make imports work
+ - DONE: cleanup Reconnector with recent changes, move generic code, make imports work
  - add SyncStatus as ObservableValue of enum in Sync
  - show load progress in LoginView/SessionPickView and do away with loading screen
  - change main.js to pass in a creation function of a SessionContainer instead of everything it is replacing
