+++ src/matrix/sessioninfo/localstorage/SessionInfoStorage.js
@@ -1,4 +1,4 @@
-export default class SessionsStore {
+export default class SessionInfoStorage {
     constructor(name) {
         this._name = name;
     }
